get rid of compiler warning.
[git-p4: depot-paths = "//depot/qt/main/": change = 165530]

+++ src/compat/tools/q3gdict.cpp
@@ -922,7 +922,7 @@ QDataStream &Q3GDict::read(QDataStream &s)
                     // but hey, serializing pointers?  can it be done
                     // at all, ever?
                     if (k)
-                        look_ptr((void *)k, d, op_insert);
+                        look_ptr((void *)(ulong)k, d, op_insert);
                 }
                 break;
         }
